requirements.txt
Django>=5.0,<6
python-dotenv>=1.0
django-ratelimit>=4.1
yookassa>=3.0

# Security: run `pip install safety && safety check` to scan for known vulnerabilities
